Allow changing main currency

+++ backend/internal/service/db_service/queries/delete_bank_account.sql
@@ -1,4 +1,4 @@
 DELETE FROM bank_accounts
 WHERE
     id = ?
-    AND organisation_id = get_current_organisation(?)
+    AND organisation_id = get_current_user_organisation_id(?)
