+++ backend/internal/service/db_service/queries/delete_employee.sql
@@ -1,4 +1,4 @@
 DELETE FROM employees
 WHERE
     id = ?
-    AND organisation_id = get_current_organisation(?)
+    AND organisation_id = get_current_user_organisation_id(?)
